KERNEL: VERSION - update kernel files to version 2.6.35.14 & improve some 2d/3d decoding files & add support for Broadcom Xscript and other clean and fix options

+++ common/arch/arm/plat-bcmap/include/plat/brcm_headset_pd.h
@@ -9,16 +9,16 @@ struct brcm_headset_pd {
 	void (*check_hs_state) (int *);
 	int hsgpio;
 	int debounce_ms;
-    short keypress_threshold_fp;
-    short keypress_threshold_lp;
-	short key_press_threshold;
-	short key_3pole_threshold;
-	short key1_threshold_l;
-	short key1_threshold_u;
-	short key2_threshold_l;
-	short key2_threshold_u;
-	short key3_threshold_l;
-	short key3_threshold_u;	
+    int keypress_threshold_fp;
+    int keypress_threshold_lp;
+	int key_press_threshold;
+	int key_3pole_threshold;
+	int key1_threshold_l;
+	int key1_threshold_u;
+	int key2_threshold_l;
+	int key2_threshold_u;
+	int key3_threshold_l;
+	int key3_threshold_u;	
 };
 
 #endif /*  __BCM_HEADSET_PD__ */
